sched/cpufreq: Add ElectroUtil cpufreq governor ElectroUtil is a schedutil based CPU Frequency governor, taking some queues from pwrutilx by @ZeroInfinity, but with an overall different frequency selection algorithm.
ElectroUtil is a hexacore-biased governor, meaning that all 4
Silver and the first 2 Gold cores are preferentially granted
higher frequencies at lower utilizations than the last 2 Gold
cores to prevent excess use of more Gold cores than necessary.

ElectroUtil is also device state aware, forcing lower
frequencies when the device is suspended. This is handled via
2 different mechanisms. The primary mechanism is the
'suspend_capacity_factor' tunable, which is used such that

max = capacity * (suspend_capacity_factor + 1) / suspend_capacity_factor

where 'max' is the divisor of the frequency calulation
algorithm. For example, a value of 10 for the capacity factor
result in a frequency that is 1/10th lower when the device is
suspended. The second mechanism is a hard limit, set by the
tuneables {gold/silver}_suspend_max_freq. If the requested
frequency is higher than than the respective suspend max, the
max will not be exceded. For the last 2 Gold cores, this max
is always the minimum frequency.

Change-Id: Id7b863c47084e4e1c3c5d0c544f6a0e97748c00c
Signed-off-by: Joe Maples <joe@frap129.org>
Signed-off-by: Atman Shah <masteratman@gmail.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>
Signed-off-by: ben863 <r.budhiono@gmail.com>